requirements.txt
# Core dependencies
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
python-dotenv>=0.19.0
sqlalchemy>=1.4.0
alembic>=1.7.0
psycopg2-binary>=2.9.0
redis>=4.0.0
rq>=1.10.0

# Document processing
PyMuPDF>=1.19.0  # For PDF processing
python-docx>=0.8.11  # For DOCX processing

# Testing
pytest>=6.2.5
pytest-asyncio>=0.15.1
pytest-cov>=2.12.1

# Development
black>=21.7b0
isort>=5.9.3
flake8>=3.9.2
mypy>=0.910
pre-commit>=2.15.0
